fix: Restore UI lint budget after main merge
Main factory Velocity/Automations pages pushed complexity and
prefer-const over budget. Split mock data, daily velocity builders,
and card mutations so check.lint.ui stays within limits.
